+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [TD-3233] Rule result remediation plan
 - [TD-4271] Support for linking implementations with business concepts
 - [TD-4621] Include data structure version `name` in GraphQL schema
+- [TD-4447] Support for associating metadata filters with a structure type
 
 ### Fixed
 
@@ -16,6 +17,11 @@
 ## Changed
 
 - [TD-4527] Exclude logically deleted sources from GraphQL response
+
+### Removed
+
+- [TD-4303] `GET /api/data_structure_types/lite` was removed, reverted
+  corresponding changes from [TD-4344]
 
 ## [4.40.0] 2022-03-14
 
@@ -149,7 +155,7 @@
 ### Added
 
 - [TD-4262] Lineage graph polling
-- [TD-4344] Add a new enpoint API for lite `structure_types` request without
+- [TD-4344] Add a new endpoint API for lite `structure_types` request without
   `metadata_fields`
 - [TD-4358] Format path in editable download CSV